app/requirements.txt
Django==4.1.4
pytz==2022.7
djangorestframework==3.14.0
django-cors-headers==3.13.0
psycopg2==2.9.5
drf-yasg==1.21.4
